Guild Modal Refactor (#268)
