+++ apps/web/src/app/[locale]/guild/detail/[id]/page.tsx
@@ -1,0 +1,5 @@
+import { GuildDetail } from '../../(components)/GuildDetail';
+
+export default function GuildDetailPage({ params }: { params: { id: string } }) {
+  return <GuildDetail guildId={params.id} />;
+}
